Move the copying of data to further outside the loop so it happens once
In SNAP, the energy groups per OpenMP thread are run sequentially.
This means that the results for lower energy groups are computed
after many calls to the octsweep subroutine.

The OpenCL memory copies of constant data are moved outside the
outer loop so they only occur once. Source data is copied for
every inner as that is were it is calculated.

We have to zero the flux edge pointers for every call of octsweep
for the reasons in paragraph 1.