Do a sweep in all 8 octants
This is done by calculating the necessary corner indexs, and then
editing the co-ordinates saved in the plane list appropiately.

Note that they are done in sequence, 0-7, lexiographically according
to their binary representation (x,y,z). This means that 4th octant
to be swept is (nx,0,0), which should be swept _after_ the 4th
octant in the case of reflective boundary conditions.