Include multiple command queues for octant sweep
The cells are spread across a number of command queues to include
some spacial parallelism. The number of queues is a compile time
constant.
All the queues need to be synchronised at the end of each
diagonal place to ensure that the edge flux data is available
for the cells in the next plane.