Use events to ensure that the planes in the sweep are executed in order
Create an event for each of the cells in the plane, waiting on all the
events in the previous planes. Call clWaitForEvents at the end of the
octant sweep to ensure that the sweep in that octant is finished.

AMD S10000: the sweep kernels in multiple queues happen concurrently
on the device :)

OS X CPU: segfault on the clWaitForEvents call :( bug report filed
OS X GPU: no segfault so far!