Remove clWaitForEvents in the octant sweep as unnecesary.
The next octant can be enqueued without waiting for this octant
to finish. This is because the last cell is enqueued on queue 0.
Because of the event list, this cell will wait until the rest of
the grid has finished. The zero_flux_buffers are then enqueued on
queue 0, before the start of the next octant sweep. The first cell
of the grid is again enqueued on queue 0, and because the queues
are IN ORDER, it is guarenteed that the kernels and memory copies
happy in the right order. This means we reduce the synchronisation.