Use a single buffer to zero the edge data and make write non-blocking
A single buffer, the biggest of the three edge arrays, is created
and is used to copy zeros into the edge buffers on the device.
This copy is also made non-blocking. This is so that the host can
continue to enqueue the next octant while the copy is taking place
thus reducing the overhead of API calls on the host side.