Use a lot of queues! Default is now set to 12.
With more queues, you seem to be able to hide some of the gaps
between kernel executions as the multiple queues can start to
overlap the scheduling of a kernel while a kernel in another
queue is still computing.

You get a much better grind time like this. For example, 12 queues on a
S9150 gets 17ns (for this commit on 16 cubed, 136 angles and 50 energy
groups, with 2 timesteps [limiting to 5 inners and 100 outers]).
Using only 4 queues this time is 14ns. This results in a solve time
(still in lockstep with the original on the CPU) of 160s to 135s.
This is a 15% reduction in time!