Use clEnqueueWaitForEvents between wavefronts instead of a wait_list
Enqueuing with a large waitlist seemed to be causing a bottleneck
in the code. This was measured by no change in run time with an
empty kernel!
This commit enqueues each kernel with no wait_list dependancies,
and enqueues a WaitForEvent on each of the command queues
containing the events of the kernels in the previous wavefront.
The current wavefront is then enqueued after this, so that the
dependancy is maintained.
On the zoo, on a K40, the runtime of a sweep has dropped from
1.46s to 0.76s.