Scalar flux (and moments) checked against epsilon not tolr
This is because the convergence tests are on epsilon, so it doesn't
make sense to check on tolr. epsilon is much less accurate than
tolr though. original snap reruns groups which are not converged to
epsilon accuracy, and so my implementation will also run all groups
which results in different tolr differences